requirements.txt
flask==3.0.2
flask-cors==4.0.0
opencv-python==4.9.0.80
google-generativeai==0.4.1
python-dotenv==1.0.1
tensorflow==2.16.1
numpy==1.26.4